requirements.txt
anndata==0.9.1
asgiref==3.7.2
autopep8==2.0.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
deprecation==2.1.0
Django==4.2.2
django-email-verification==0.3.1
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
llvmlite==0.40.1rc1
mysqlclient==2.1.1
natsort==8.3.1
numba==0.57.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pycodestyle==2.10.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
zipp==3.15.0
